requirements.txt
streamlit==1.10.0
pydantic==1.8.2
unstructured==0.1.2
langchain-core==0.0.10
langchain-openai==0.0.10
langchain-community==0.0.10
langchain-chroma==0.0.10
langchain-retrievers==0.0.10
langchain-storage==0.0.10
python-dotenv==0.19.2
requests==2.26.0
pypdf2==1.26.0
pdfminer.six
pillow-heif
opencv-python
pdf2image==1.17.0
unstructured[all-docs]
unstructured_inference
unstructured_pytesseract
python-poppler
poppler-utils
pytesseract
nltk==3.8.1
base64
uuid
